Fix signed and unsigned integer comparison warning

+++ src/transport/tcp.c
@@ -60,7 +60,7 @@ static int SmolRTSP_TcpTransport_transmit(VSelf, SmolRTSP_IoVecSlice bufs) {
     for (size_t i = 0; i < bufs.len; i++) {
         const CharSlice99 vec = CharSlice99_new(bufs.ptr[i].iov_base, bufs.ptr[i].iov_len);
         ret = VCALL(self->w, write, vec);
-        if (ret != vec.len) {
+        if (ret != (ssize_t)vec.len) {
             return -1;
         }
     }
